vscode-demos: libgpiod-c: Add multi-threaded button support to demo
Add button framework support and apply to SOM_MX8M_MINI platform alone
for now. This uses another thread to read any of the push buttons on the
symphony board and increments the count value. The value is normally
decremented by the LED blinking routine, so pressing the button extends
the lifetime of the application run.

Signed-off-by: Ken sloat <ken.s@variscite.com>