vscode-demos: libgpiod-c: Improve vscode settings for build/debug
Utilize existing SDK variables which helps to simplify project settings
and those which need to be modified. Additionally, utilize GDB in the
SDK which will have the best chance of compatibility and not require
user to install another tool.

Using this setup requires sourcing the Yocto SDK before code is
launched.

Signed-off-by: Ken sloat <ken.s@variscite.com>